requirements.txt
absl-py==0.12.0
astor==0.8.1
cached-property==1.5.2
certifi==2020.12.5
cffi==1.14.5
cloudpickle==1.6.0
colorama==0.4.4
Cython==0.29.22
dataclasses==0.8
fasteners==0.16
future==0.18.2
gast==0.2.2
glfw==2.1.0
google-pasta==0.2.0
grpcio==1.36.1
gym==0.18.0
h5py==3.1.0
imageio==2.9.0
importlib-metadata==3.10.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.4
memory-profiler==0.58.0
mujoco-py==2.0.2.13
nn==0.1.1
numexpr==2.7.3
numpy==1.19.5
opt-einsum==3.3.0
passer==0.0.1
Pillow==7.2.0
protobuf==3.15.6
psutil==5.8.0
pycparser==2.20
pyglet==1.5.0
PyYAML==5.4.1
scipy==1.5.4
six==1.15.0
tables==3.6.1
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
Theano==1.0.5
torch==1.7.1+cu101
torchaudio==0.7.2
torchvision==0.8.2+cu101
typing-extensions==3.7.4.3
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.1
